extract default endpoints into namespace
