+++ lib/api/v3/work_packages/create_form_api.rb
@@ -31,11 +31,12 @@ module API
     module WorkPackages
       class CreateFormAPI < ::API::OpenProjectAPI
         resource :form do
-          post &::API::V3::Utilities::DefaultCreateForm.new(model: WorkPackage,
-                                                            parse_service: WorkPackages::ParseParamsService,
-                                                            instance_generator: ->(*) {
-                                                              WorkPackage.new(author: current_user)
-                                                            }).mount
+          post &::API::V3::Utilities::Endpoints::CreateForm.new(model: WorkPackage,
+                                                                parse_service: WorkPackages::ParseParamsService,
+                                                                instance_generator: ->(*) {
+                                                                  WorkPackage.new(author: current_user)
+                                                                })
+                                                           .mount
         end
       end
     end
