+++ lib/api/v3/work_packages/update_end_point.rb
@@ -1,7 +1,7 @@
 module API
   module V3
     module WorkPackages
-      class UpdateEndPoint < API::V3::Utilities::DefaultUpdate
+      class UpdateEndPoint < API::V3::Utilities::Endpoints::Update
         def present_success(current_user, call)
           call.result.reload
