+++ modules/grids/app/controllers/api/v3/grids/create_form_api.rb
@@ -31,10 +31,11 @@ module API
     module Grids
       class CreateFormAPI < ::API::OpenProjectAPI
         resource :form do
-          post &::API::V3::Utilities::DefaultCreateForm.new(model: ::Grids::Grid,
-                                                            instance_generator: ->(params) {
-                                                              ::Grids::Factory.build(params.delete(:scope), current_user)
-                                                            }).mount
+          post &::API::V3::Utilities::Endpoints::CreateForm.new(model: ::Grids::Grid,
+                                                                instance_generator: ->(params) {
+                                                                  ::Grids::Factory.build(params.delete(:scope), current_user)
+                                                                })
+                                                           .mount
         end
       end
     end
